requirements.txt
amqp==5.0.5
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.3.0
boto==2.49.0
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-apptemplates==1.5
django-contrib-comments==2.0.0
django-cors-headers==3.7.0
django-js-reverse==0.9.1
django-queryset-csv==1.1.0
django-recaptcha==2.0.6
django-redis==4.12.1
django-registration-redux==2.9
django-storages==1.11.1
django-threadedcomments==1.2
django-tinsel==1.0.2
django-webpack-loader==0.7.0
flake8==3.8.4
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jsonschema==3.2.0
kombu==5.0.2
libsass==0.20.1
lockfile==0.12.2
mccabe==0.6.1
modgrammar==0.10
msgpack==0.6.2
olefile==0.46
packaging==20.3
pep517==0.8.2
pep8==1.7.1
Pillow==8.1.0
progress==1.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-omgeo==6.0.4
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
rollbar==0.15.2
six==1.14.0
sqlparse==0.4.1
unicodecsv==0.14.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
